+++ src/Exception/BadFunctionCallException.php
@@ -28,5 +28,5 @@ namespace Inane\Stdlib\Exception;
  * @implements \Inane\Stdlib\Exception\ExceptionInterface
  * @version 0.2.0
  */
-class BadFunctionCallException extends \BadFunctionCallException implements ExceptionInterface {
+class BadFunctionCallException extends LogicException {
 }