+++ src/Exception/InvalidArgumentException.php
@@ -21,8 +21,6 @@ declare(strict_types=1);
 
 namespace Inane\Stdlib\Exception;
 
-use Inane\Stdlib\Exception\LogicException;
-
 /**
  * Exception thrown if an argument is not of the expected type.
  *
@@ -30,7 +28,7 @@ use Inane\Stdlib\Exception\LogicException;
  * @implements \Inane\Stdlib\Exception\ExceptionInterface
  * @version 0.3.0
  */
-class InvalidArgumentException extends LogicException implements ExceptionInterface {
+class InvalidArgumentException extends LogicException {
     // base code for invalid argument exceptions
-    protected $code = 750;
+    protected $code = 355;
 }