+++ src/Exception/LengthException.php
@@ -28,5 +28,5 @@ namespace Inane\Stdlib\Exception;
  * @implements \Inane\Stdlib\Exception\ExceptionInterface
  * @version 0.2.0
  */
-class LengthException extends \LengthException implements ExceptionInterface {
+class LengthException extends LogicException {
 }