+++ src/Exception/OverflowException.php
@@ -28,5 +28,5 @@ namespace Inane\Stdlib\Exception;
  * @implements \Inane\Stdlib\Exception\ExceptionInterface
  * @version 0.2.0
  */
-class OverflowException extends \OverflowException implements ExceptionInterface {
+class OverflowException extends RuntimeException {
 }