+++ src/Exception/UnexpectedValueException.php
@@ -28,5 +28,5 @@ namespace Inane\Stdlib\Exception;
  * @implements \Inane\Stdlib\Exception\ExceptionInterface
  * @version 0.2.0
  */
-class UnexpectedValueException extends \UnexpectedValueException implements ExceptionInterface {
+class UnexpectedValueException extends RuntimeException {
 }